Move plan functions to own file
